Critical Bug Fix
Found a bug where the hashes being generated between randomx insances were not consistent even though the inputs were the same. Fixed it and added an additional test for this.

Modified crate to accept &[u8] instead of &[str], as well as to return Vec[u8] instead of String, to be consistent with other hashing crates.

Updated version since this is a breaking change.